Tweaking CatchMissingRefreshToken a little more.

+++ src/Actions/Sync/CatchMissingRefreshTokenAction.php
@@ -17,16 +17,23 @@ class CatchMissingRefreshTokenAction
 
     protected $roles_to_temporary_remove;
 
+    protected $has_missing_refresh_token;
+
+    protected $superuser_roles;
+
     public function __construct()
     {
         $this->get_main_character_action = new GetMainCharacterAction;
         $this->roles_to_temporary_remove = collect();
+        $this->superuser_roles = collect();
+
     }
 
     public function execute(Group $group)
     {
         $this->group = $group;
         $main_character = $this->get_main_character_action->execute($this->group);
+        $this->has_missing_refresh_token = false;
 
         foreach ($this->group->users as $user) {
 
@@ -39,30 +46,40 @@ class CatchMissingRefreshTokenAction
 
                 // Throw missing refresh token event
                 event(new MissingRefreshToken($user, $main_character));
+
+                $this->has_missing_refresh_token = true;
             }
+
         }
 
-        SeatGroup::all()->each(function ($seatgroup) {
+        if($this->has_missing_refresh_token) {
+            SeatGroup::all()->each(function ($seatgroup) {
 
-            // If group is member and no longer qualified
-            if (!$seatgroup->isQualified($this->group) && $seatgroup->isMember($this->group)) {
+                // If group is member and no longer qualified
+                if (!$seatgroup->isQualified($this->group) && $seatgroup->isMember($this->group)) {
 
-                // remove member status
-                $seatgroup->member()->detach($this->group->id);
-            }
+                    // remove member status
+                    $seatgroup->member()->detach($this->group->id);
+                }
 
-            // if group is member and still qualified
-            if ($seatgroup->isQualified($this->group) && $seatgroup->isMember($this->group)) {
+                // if group is member and still qualified
+                if ($seatgroup->isQualified($this->group) && $seatgroup->isMember($this->group)) {
 
-                // add role id's to roles_to_temporary_remove
-                $seatgroup->role->each(function ($role) {
-                    $this->roles_to_temporary_remove->push($role->id);
-                });
-            }
+                    // add role id's to roles_to_temporary_remove
+                    $seatgroup->role->each(function ($role) {
+                        $this->roles_to_temporary_remove->push($role->id);
+                    });
+                }
 
-        });
+            });
+        }
 
-        return $this->roles_to_temporary_remove;
+        $this->superuser_roles = (new CatchSuperuserAction)->execute($this->group);
+
+        return $this->roles_to_temporary_remove
+            ->reject(function ($role) {
+                return in_array($role, $this->superuser_roles->toArray());
+            });
     }
 
 }
